Mechanisms for showing input form and adding new lists, items works!
Some example of things I still need to feed include clearing the input
fields as well as the input container on each new click. Also need to
make sure that the current "Counter" system for tracking which list to
add the items to will scale. Highly doubtful though. Also have a lot of
formatting work ahead!